Now passes flags to ua.ahk and supports new errors

+++ maintenance/te.ahk
@@ -1,25 +1,50 @@
-sleep 500
-computerbegin:
-sleep 500
+;;;; te.ahk v2
+;;;; The idea is, I test a computer for TestNav compatibility and try to fix any issues I find.
+ranUAatLeastOnce = false
 runwait QRes.exe /x 1280 /y 1024
 sleep 500
 goto,testbegin
+
 msgbox,48,error,Error! This line should never be seen
+
+;;ua begin
 uabegin:
 sleep 5000
+;; Keep UA from restarting when done
+if A_OSVersion in WIN_7,WIN_VISTA
+{
+	RunWait %ComSpec% /C "echo foo > %HOMEDRIVE%%HOMEPATH%\Desktop\ahknorestart.txt"
+}
+if A_OSVersion in WIN_XP
+{
+	RunWait %ComSpec% /C "echo foo > "C:\Documents and Settings\student\Desktop\ahknorestart.txt""
+}
 SetTimer,altr,-10000
 runwait ua.exe
+ranUAatLeastOnce = true
 sleep 5000
 goto,testbegin
+
 msgbox,48,error,Error! This line should never be seen
+
+;;test begin
 testbegin:
 Run %ComSpec% /C ""C:\Program Files\Internet Explorer\iexplore.exe" https://proctorcaching.pearsonaccess.com/ems/systemCheck/systemCheck.jsp?acc=tx"
 sleep 15000
 IfWinExist,Java Update Needed
 {
-	Run %ComSpec% /C "taskkill /F /IM iexplore.exe"
-	msgbox,48,Fail,It seems java is not up to date. I'm going to try to update it now.,5
-	goto,uabegin
+	if ranUAatLeastOnce in false
+	{
+		Run %ComSpec% /C "taskkill /F /IM iexplore.exe"
+		msgbox,0,Fail,It seems java is not up to date. I'm going to try to update it now.,5
+		goto,uabegin
+	}
+	if ranUAatLeastOnce in true
+	{
+		msgbox,48,Picnic!,Well I've already tried updating it once so I won't try again. You'll have to test this thing thing manually.
+		Run %ComSpec% /C ""C:\Program Files\Internet Explorer\iexplore.exe" http://tx.testnav.com/txqc"
+		exitapp
+	}
 }
 sleep 5000
 IfWinExist,Security Warning
@@ -43,6 +68,17 @@ IfWinExist,Internet Explorer 9 - Microsoft Windows - Windows Internet Explorer
 	WinActivate
 	sleep 500
 	send ^w
+}
+sleep 5000
+loop,2
+{
+	IfWinExist,Security Information
+	{
+		WinActivate
+		sleep 500
+		send {enter}
+	}
+	sleep 10000
 }
 sleep 2000
 IfWinExist,System Check for TestNav - Windows Internet Explorer
@@ -92,7 +128,7 @@ IfWinExist,Security Warning
 }
 sleep 5000
 IfWinExist,TestNav - Windows Internet Explorer
-WinActivate
+	WinActivate
 sleep 500
 send {F11}
 sleep 2000
